+++ ledger/views/index.py
@@ -9,7 +9,7 @@ from django.shortcuts import redirect
 
 # AA Ledger
 # Ledger
-from ledger.view_helpers.core import add_info_to_context
+from ledger.helpers.core import add_info_to_context
 
 logger = logging.getLogger(__name__)